pyproject.toml
[project]
name = "news-article"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-cognitiveservices-speech>=1.48.1",
    "dotenv>=0.9.9",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.9",
    "langgraph>=1.0.8",
    "moviepy>=2.2.1",
    "openai>=2.20.0",
    "typer>=0.21.2",
]
